pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "flask>=3.1.2",
    "jinja2>=3.1.6",
    "pefile>=2024.8.26",
    "pygments>=2.19.2",
    "yara-python>=4.5.4",
]
